fix: render univer.js only in browser
